requirements_dev.txt
-r requirements.txt
flake8==3.8.3
requests-mock==1.8.0
coverage==5.2
